requirements.txt
flask>=2.0.0
requests>=2.25.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
